+++ dialects/mssql/mssql.go
@@ -1,0 +1,3 @@
+package mssql
+
+import _ "github.com/denisenkom/go-mssqldb"