+++ dialects/sqlite/sqlite.go
@@ -1,0 +1,3 @@
+package sqlite
+
+import _ "github.com/mattn/go-sqlite3"